backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
alembic>=1.13.0
python-dotenv>=1.0.0

# Security Dependencies
slowapi>=0.1.9
cryptography>=41.0.0
bcrypt>=4.0.1
email-validator>=2.0.0

# Input Validation and Sanitization
bleach>=6.0.0
markupsafe>=2.1.0

# Security Headers
secure>=0.3.0

# Monitoring and Logging
structlog>=23.1.0
sentry-sdk[fastapi]>=1.32.0

# Additional Security
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# (AI removed)
